no-mistakes(review): latch rating prompt state, share detail harness, refresh docs
