+++ AscendAppTests/LiveClimbHeartRateChartSnapshotTests.swift
